+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# Unreleased
+- Added `hogan/module/grid/theme` filter. With this you can now add section theme classname.
+
 ## 1.1.5
 - Fixed a bug on large cards when text align is center. Meta/icon was left align. They are now center aligned. [#32](https://github.com/DekodeInteraktiv/hogan-grid/issues/32)